# api/openapi.yaml
openapi: 3.0.3
info:
  title: Diary - OpenAPI 3.0
  contact:
    email: ilya.korolev@outlook.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1

paths:
  /v1/authorize:
    post:
      tags:
        - auth
      summary: validate user/password and return token
      security: [] # Override to indicate no security required
      operationId: authorize
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthData"
        required: true
      responses:
        "200":
          description: return token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "JWT token"
                required:
                  - token
        "401":
          description: Authentication failed

  /v1/user:
    get:
      tags:
        - user
      summary: return user object
      operationId: getUser
      responses:
        "200":
          description: user object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /v1/assets:
    get:
      tags:
        - assets
      summary: return asset by path
      operationId: getAsset
      parameters:
        - name: path
          in: query
          description: relative path to asset file
          required: true
          schema:
            type: string
          example: "images/photos/vacation.jpg"
      responses:
        "200":
          description: return asset
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Asset not found

  /v1/assets/batch:
    post:
      tags:
        - assets
      summary: upload multiple asset files
      operationId: uploadAssetsBatch
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                assets:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: "Asset files to upload"
              required:
                - assets
        required: true
      responses:
        "200":
          description: assets uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetsBatchResponse"
        "400":
          description: Bad request - invalid files or missing assets field
        "401":
          description: Unauthorized - authentication required
        "413":
          description: Payload too large - batch or file size limits exceeded
        "500":
          description: Internal server error - failed to save files

  /v1/items:
    get:
      tags:
        - items
      summary: get diary items
      operationId: getItems
      parameters:
        - name: date
          in: query
          description: filter items by date (optional)
          required: false
          schema:
            type: string
            format: date
          example: "2024-01-15"
        - name: search
          in: query
          description: search text to filter items by title and body content
          required: false
          schema:
            type: string
          example: "vacation"
        - name: tags
          in: query
          description: comma-separated list of tags to filter items
          required: false
          schema:
            type: string
          example: "personal,work"
      responses:
        "200":
          description: diary items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemsListResponse"
        "400":
          description: Invalid parameters
        "404":
          description: No items found
    put:
      tags:
        - items
      summary: upsert diary item
      operationId: putItems
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemsRequest"
        required: true
      responses:
        "200":
          description: item saved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemsResponse"
        "400":
          description: Invalid request data
        "401":
          description: Unauthorized

  /v1/health/issues:
    get:
      tags:
        - health
      summary: get storage health issues for current user
      operationId: getHealthIssues
      responses:
        "200":
          description: health issues for the authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthIssuesResponse"
        "401":
          description: Unauthorized

  /v1/health/fix:
    post:
      tags:
        - health
      summary: fix storage issues for current user
      operationId: fixHealthIssues
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HealthFixRequest"
        required: true
      responses:
        "200":
          description: updated health issues after applying fixes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthIssuesResponse"
        "400":
          description: Invalid request body
        "401":
          description: Unauthorized
        "500":
          description: Fix failed

  /v1/sync/changes:
    get:
      tags:
        - sync
      summary: get changes for synchronization
      operationId: getChanges
      parameters:
        - name: since
          in: query
          description: get changes since this change ID (exclusive)
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
          example: 123
        - name: limit
          in: query
          description: maximum number of changes to return
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 1000
            default: 100
          example: 50
      responses:
        "200":
          description: changes retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResponse"
        "400":
          description: Invalid parameters
        "401":
          description: Unauthorized

security:
  - BearerAuth: []

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthData:
      type: object
      properties:
        email:
          type: string
          example: john@email.com
        password:
          type: string
          example: "12345"
      required:
        - email
        - password

    Entity:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
      required:
        - id

    User:
      allOf:
        - $ref: "#/components/schemas/Entity"
        - type: object
          properties:
            email:
              type: string
            startDate:
              type: string
              format: date-time
          required:
            - email
            - startDate

    ItemsRequest:
      type: object
      properties:
        date:
          type: string
          format: date
          example: "2024-01-15"
        title:
          type: string
          example: "My diary entry"
        tags:
          type: array
          items:
            type: string
          example: ["personal", "work", "travel"]
        body:
          type: string
          example: "Today was a great day..."
      required:
        - date
        - title

    ItemsResponse:
      type: object
      properties:
        date:
          type: string
          format: date
          example: "2024-01-15"
        title:
          type: string
          example: "My diary entry"
        tags:
          type: array
          items:
            type: string
          example: ["personal", "work", "travel"]
        body:
          type: string
          example: "Today was a great day..."
        previousDate:
          type: string
          format: date
          nullable: true
          example: "2024-01-14"
        nextDate:
          type: string
          format: date
          nullable: true
          example: "2024-01-16"
      required:
        - date
        - title

    ItemsListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ItemsResponse"
          description: "List of diary items matching the search criteria"
        totalCount:
          type: integer
          description: "Total number of items found"
          example: 42
      required:
        - items
        - totalCount

    SyncResponse:
      type: object
      properties:
        changes:
          type: array
          items:
            $ref: "#/components/schemas/SyncChangeResponse"
          description: "List of changes since the requested ID"
        hasMore:
          type: boolean
          description: "Whether there are more changes available"
          example: true
        nextId:
          type: integer
          format: int32
          description: "ID to use for the next sync request (if hasMore is true)"
          example: 456
      required:
        - changes
        - hasMore

    SyncChangeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: "Unique change ID"
          example: 123
        userId:
          type: string
          description: "User ID who made the change"
          example: "user-123"
        date:
          type: string
          format: date
          description: "Date of the diary entry that was changed"
          example: "2024-01-15"
        operationType:
          type: string
          enum: ["created", "updated", "deleted"]
          description: "Type of operation performed"
          example: "updated"
        timestamp:
          type: string
          format: date-time
          description: "When the change occurred"
          example: "2024-01-15T10:30:00Z"
        itemSnapshot:
          allOf:
            - $ref: "#/components/schemas/ItemsResponse"
          nullable: true
          description: "Current state of the item (null for deleted items)"
        metadata:
          type: array
          items:
            type: string
          description: "Additional metadata about the change"
          example: ["mobile-app", "v1.2.3"]
      required:
        - id
        - userId
        - date
        - operationType
        - timestamp

    HealthIssue:
      type: object
      properties:
        check:
          type: string
          description: Name of the check that found this issue
          example: "mime"
        path:
          type: string
          description: Filesystem path of the affected asset
          example: "/data/diary-assets/user-id/abc123.jpg"
        message:
          type: string
          description: Human-readable description of the issue
          example: "video file saved as .jpg, should be .mp4"
        fixable:
          type: boolean
          description: Whether an automated fix is available
          example: true
      required:
        - check
        - path
        - message
        - fixable

    HealthIssuesResponse:
      type: object
      properties:
        issues:
          type: array
          items:
            $ref: "#/components/schemas/HealthIssue"
          description: List of storage issues found for this user
        lastChecked:
          type: string
          format: date-time
          nullable: true
          description: When the last health check ran (null if never run)
          example: "2026-04-01T12:00:00Z"
      required:
        - issues

    HealthFixRequest:
      type: object
      properties:
        checks:
          type: array
          items:
            type: string
          description: Names of checks to run with fix=true. Empty array runs all checks.
          example: ["mime"]
      required:
        - checks

    AssetsBatchFile:
      type: object
      properties:
        originalName:
          type: string
          description: Original filename provided by the client
          example: "photo.jpg"
        savedName:
          type: string
          description: Server-side stored filename
          example: "123e4567-e89b-12d3-a456-426614174000.jpg"
        size:
          type: integer
          format: int64
          description: File size in bytes
          example: 102400
        contentType:
          type: string
          description: MIME type detected for the file
          example: "image/jpeg"
      required:
        - originalName
        - savedName
        - size

    AssetsBatchResponse:
      type: object
      properties:
        files:
          type: array
          items:
            $ref: "#/components/schemas/AssetsBatchFile"
        count:
          type: integer
          description: Number of files successfully uploaded
          example: 3
      required:
        - files
        - count